+++ src/main/kotlin/ar/edu/unq/desapp/grupoL/criptop2p/webservice/UserRestService.kt
@@ -71,7 +71,7 @@ class UserRestService {
     fun userById(@PathVariable("id") id: Int): ResponseEntity<*> {
         var response : ResponseEntity<*>?
         try {
-            val newUser = userService.findByID(id)
+            val newUser = userService.findByID(id.toLong())
             val userView =   UserViewMapper(newUser.id, newUser.name, newUser.surname, newUser.email, newUser.address, newUser.cvu, newUser.walletAddress)
 
             ResponseEntity.status(200)
@@ -92,7 +92,7 @@ class UserRestService {
     fun update (@PathVariable("id") id: Int,@RequestBody entity: UserUpdateMapper): ResponseEntity<*> {
         var response : ResponseEntity<*>?
         try {
-            val userview = userService.update(id,entity)
+            val userview = userService.update(id.toLong(),entity)
 
             ResponseEntity.status(200)
             response = ResponseEntity.ok().body(userview)
@@ -111,7 +111,7 @@ class UserRestService {
     fun deleteUserById(@PathVariable("id") id: Int): ResponseEntity<*> {
         var response : ResponseEntity<*>?
         try {
-            userService!!.deleteById(id)
+            userService!!.deleteById(id.toLong())
             val resultado: MutableMap<String, Int> = HashMap()
             resultado["succesfully user deleted with iD"] = id
             response = ResponseEntity.ok().body<Map<String, Int>>(resultado)
